Relax query tolerances. Exact transformation is Proj version dependent.
In some cases there are multiple options for transforming from one CRS to another.
For EPSG:4326 (WGS84) to EPSG:3311 (UTM zone 11) there are two options in Proj 6.3 to 7.2
(use projinfo to see the options), one that does a shift for WGS84 to GRS80 and one that does not.
Proj 6.3 and 7.1 use the transformation without the shift and Proj 7.2 uses the one with the shift.
This leads to slight differences (~1m) in query locations. The best workaround is to relax the tolerances
used in the query checks from 1.0e-6 to 1.0e-5.